requirements/common.txt
Django==1.11.3
django-custom-user==0.7
dj-database-url==0.4.2
psycopg2==2.7.1
asn1crypto==0.22.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cryptography==2.0
django-cors-headers==2.1.0
django-rest-knox==3.0.0
djangorestframework==3.6.3
idna==2.5
pycparser==2.18
pyOpenSSL==17.1.0
pytz==2017.2
requests==2.18.3
six==1.10.0
urllib3==1.22
